Add namespacing and new directory structure to robot-name
