--- exercises/robot-name/robot-name.php
@@ -1,1 +1,0 @@
-<?php
