+++ exercises/robot-name/src/Robot.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Exercism\RobotName;
+
+class Robot
+{
+}
